MC farming bot
Put the crop in your hand, set the horizontal size of your farm, and line your player up so that their face is parallel to the farm. Along with that move your cursor to the crop you wish to plant and destroy.

PS. For this to work you must remap your destroy button to (x) in MC settings.